Module google_retail2::api

Structs

CloudRetail
Central instance to access all CloudRetail related resource activities

GoogleApiHttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page. This message can be used both in streaming and non-streaming API methods in the request as well as the response. It can be used as a top-level request field, which is convenient if one wants to extract parameters from either the URL or HTTP template into the request fields and also want access to the raw HTTP body. Example: message GetResourceRequest { // A unique request id. string request_id = 1; // The raw HTTP body is bound to this field. google.api.HttpBody http_body = 2; } service ResourceService { rpc GetResource(GetResourceRequest) returns (google.api.HttpBody); rpc UpdateResource(google.api.HttpBody) returns (google.protobuf.Empty); } Example with streaming methods: service CaldavService { rpc GetCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); rpc UpdateCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); } Use of this type only changes how the request and response bodies are handled, all other features will continue to work unchanged.

GoogleCloudRetailV2BigQuerySource
BigQuery source import data from.

GoogleCloudRetailV2Catalog
The catalog configuration.

GoogleCloudRetailV2CustomAttribute
A custom attribute that is not explicitly modeled in Product.

GoogleCloudRetailV2GcsSource
Google Cloud Storage location for input content. format.

GoogleCloudRetailV2Image
Product thumbnail/detail image.

GoogleCloudRetailV2ImportErrorsConfig
Configuration of destination for Import related errors.

GoogleCloudRetailV2ImportProductsRequest
Request message for Import methods.

GoogleCloudRetailV2ImportUserEventsRequest
Request message for the ImportUserEvents request.

GoogleCloudRetailV2ListCatalogsResponse
Response for CatalogService.ListCatalogs method.

GoogleCloudRetailV2PredictRequest
Request message for Predict method.

GoogleCloudRetailV2PredictResponse
Response message for predict method.

GoogleCloudRetailV2PredictResponsePredictionResult
PredictionResult represents the recommendation prediction results.

GoogleCloudRetailV2PriceInfo
The price information of a Product.

GoogleCloudRetailV2Product
Product captures all metadata information of items to be recommended or searched.

GoogleCloudRetailV2ProductDetail
Detailed product information associated with a user event.

GoogleCloudRetailV2ProductInlineSource
The inline source for the input config for ImportProducts method.

GoogleCloudRetailV2ProductInputConfig
The input config source for products.

GoogleCloudRetailV2ProductLevelConfig
Configures what level the product should be uploaded with regards to how users will be send events and how predictions will be made.

GoogleCloudRetailV2PurchaseTransaction
A transaction represents the entire purchase transaction.

GoogleCloudRetailV2PurgeUserEventsRequest
Request message for PurgeUserEvents method.

GoogleCloudRetailV2RejoinUserEventsRequest
Request message for RejoinUserEvents method.

GoogleCloudRetailV2UserEvent
UserEvent captures all metadata information Retail API needs to know about how end users interact with customers' website.

GoogleCloudRetailV2UserEventInlineSource
The inline source for the input config for ImportUserEvents method.

GoogleCloudRetailV2UserEventInputConfig
The input config source for user events.

GoogleCloudRetailV2UserInfo
Information of an end user.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

ProjectLocationCatalogBrancheOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationCatalogBrancheProductCreateCall
Creates a Product.

ProjectLocationCatalogBrancheProductDeleteCall
Deletes a Product.

ProjectLocationCatalogBrancheProductGetCall
Gets a Product.

ProjectLocationCatalogBrancheProductImportCall
Bulk import of multiple Products. Request processing may be synchronous. No partial updating is supported. Non-existing items are created. Note that it is possible for a subset of the Products to be successfully updated.

ProjectLocationCatalogBrancheProductPatchCall
Updates a Product.

ProjectLocationCatalogListCall
Lists all the Catalogs associated with the project.

ProjectLocationCatalogOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationCatalogOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationCatalogPatchCall
Updates the Catalogs.

ProjectLocationCatalogPlacementPredictCall
Makes a recommendation prediction.

ProjectLocationCatalogUserEventCollectCall
Writes a single user event from the browser. This uses a GET request to due to browser restriction of POST-ing to a 3rd party domain. This method is used only by the Retail API JavaScript pixel and Google Tag Manager. Users should not call this method directly.

ProjectLocationCatalogUserEventImportCall
Bulk import of User events. Request processing might be synchronous. Events that already exist are skipped. Use this method for backfilling historical user events. Operation.response is of type ImportResponse. Note that it is possible for a subset of the items to be successfully inserted. Operation.metadata is of type ImportMetadata.

ProjectLocationCatalogUserEventPurgeCall
Deletes permanently all user events specified by the filter provided. Depending on the number of events specified by the filter, this operation could take hours or days to complete. To test a filter, use the list command first.

ProjectLocationCatalogUserEventRejoinCall
Triggers a user event rejoin operation with latest product catalog. Events will not be annotated with detailed product information if product is missing from the catalog at the time the user event is ingested, and these events are stored as unjoined events with a limited usage on training and serving. This API can be used to trigger a 'join' operation on specified events with latest version of product catalog. It can also be used to correct events joined with wrong product catalog.

ProjectLocationCatalogUserEventWriteCall
Writes a single user event.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudRetail hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.