Struct google_retail2::api::GoogleCloudRetailV2UserEvent

pub struct GoogleCloudRetailV2UserEvent {
    pub attributes: Option<HashMap<String, GoogleCloudRetailV2CustomAttribute>>,
    pub attribution_token: Option<String>,
    pub cart_id: Option<String>,
    pub event_time: Option<String>,
    pub event_type: Option<String>,
    pub experiment_ids: Option<Vec<String>>,
    pub page_categories: Option<Vec<String>>,
    pub page_view_id: Option<String>,
    pub product_details: Option<Vec<GoogleCloudRetailV2ProductDetail>>,
    pub purchase_transaction: Option<GoogleCloudRetailV2PurchaseTransaction>,
    pub referrer_uri: Option<String>,
    pub search_query: Option<String>,
    pub uri: Option<String>,
    pub user_info: Option<GoogleCloudRetailV2UserInfo>,
    pub visitor_id: Option<String>,
}

UserEvent captures all metadata information Retail API needs to know about how end users interact with customers' website.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations catalogs user events write projects (request|response)

Fields
attributes: Option<HashMap<String, GoogleCloudRetailV2CustomAttribute>>

Extra user event features to include in the recommendation model. The key must be a UTF-8 encoded string with a length limit of 5,000 characters. Otherwise, an INVALID_ARGUMENT error is returned. For product recommendation, an example of extra user information is traffic_channel, i.e. how user arrives at the site. Users can arrive at the site by coming to the site directly, or coming through Google search, and etc.

attribution_token: Option<String>

Highly recommended for user events that are the result of PredictionService.Predict. This field enables accurate attribution of recommendation model performance. The value must be a valid PredictResponse.attribution_token for user events that are the result of PredictionService.Predict. This token enables us to accurately attribute page view or purchase back to the event and the particular predict response containing this clicked/purchased product. If user clicks on product K in the recommendation results, pass PredictResponse.attribution_token as a URL parameter to product K's page. When recording events on product K's page, log the PredictResponse.attribution_token to this field.

cart_id: Option<String>

The id or name of the associated shopping cart. This id is used to associate multiple items added or present in the cart before purchase. This can only be set for add-to-cart, purchase-complete, or shopping-cart-page-view events.

event_time: Option<String>

Only required for UserEventService.ImportUserEvents method. Timestamp of when the user event happened.

event_type: Option<String>

Required. User event type. Allowed values are: * add-to-cart: Products being added to cart. * category-page-view: Special pages such as sale or promotion pages viewed. * detail-page-view: Products detail page viewed. * home-page-view: Homepage viewed. * promotion-offered: Promotion is offered to a user. * promotion-not-offered: Promotion is not offered to a user. * purchase-complete: User finishing a purchase. * search: Product search. * shopping-cart-page-view: User viewing a shopping cart.

experiment_ids: Option<Vec<String>>

A list of identifiers for the independent experiment groups this user event belongs to. This is used to distinguish between user events associated with different experiment setups (e.g. using Retail API, using different recommendation models).

page_categories: Option<Vec<String>>

The categories associated with a category page. To represent full path of category, use '>' sign to separate different hierarchies. If '>' is part of the category name, please replace it with other character(s). Category pages include special pages such as sales or promotions. For instance, a special sale page may have the category hierarchy: "pageCategories" : ["Sales > 2017 Black Friday Deals"]. Required for category-page-view events. At least one of search_query or page_categories is required for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

page_view_id: Option<String>

A unique id of a web page view. This should be kept the same for all user events triggered from the same pageview. For example, an item detail page view could trigger multiple events as the user is browsing the page. The pageViewId property should be kept the same for all these events so that they can be grouped together properly. When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically.

product_details: Option<Vec<GoogleCloudRetailV2ProductDetail>>

The main product details related to the event. This field is required for the following event types: * add-to-cart * detail-page-view * purchase-complete In a search event, this field represents the products returned to the end user on the current page (the end user may have not finished broswing the whole page yet). When a new page is returned to the end user, after pagination/filtering/ordering even for the same query, a new search event with different product_details is desired. The end user may have not finished broswing the whole page yet.

purchase_transaction: Option<GoogleCloudRetailV2PurchaseTransaction>

A transaction represents the entire purchase transaction. Required for purchase-complete events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

referrer_uri: Option<String>

The referrer URL of the current page. When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically.

search_query: Option<String>

The user's search query. The value must be a UTF-8 encoded string with a length limit of 5,000 characters. Otherwise, an INVALID_ARGUMENT error is returned. At least one of search_query or page_categories is required for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

uri: Option<String>

Complete URL (window.location.href) of the user's current page. When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically. Maximum length 5,000 characters.

user_info: Option<GoogleCloudRetailV2UserInfo>

User information.

visitor_id: Option<String>

Required. A unique identifier for tracking visitors. For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor log in/out of the website. The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.