Struct google_retail2::api::ProjectLocationCatalogBrancheProductImportCall

pub struct ProjectLocationCatalogBrancheProductImportCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Bulk import of multiple Products. Request processing may be synchronous. No partial updating is supported. Non-existing items are created. Note that it is possible for a subset of the Products to be successfully updated.

A builder for the locations.catalogs.branches.products.import method supported by a project resource. It is not used directly, but through a ProjectMethods instance.

Example

Instantiate a resource method builder

use retail2::api::GoogleCloudRetailV2ImportProductsRequest;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = GoogleCloudRetailV2ImportProductsRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().locations_catalogs_branches_products_import(req, "parent")
             .doit().await;

Implementations

impl<'a, C> ProjectLocationCatalogBrancheProductImportCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, GoogleLongrunningOperation)>

Perform the operation you have build so far.

pub fn request(
self,
new_value: GoogleCloudRetailV2ImportProductsRequest
) -> ProjectLocationCatalogBrancheProductImportCall<'a, C>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn parent(
self,
new_value: &str
) -> ProjectLocationCatalogBrancheProductImportCall<'a, C>

Required. projects/1234/locations/global/catalogs/default_catalog/branches/default_branch If no updateMask is specified, requires products.create permission. If updateMask is specified, requires products.update permission.

Sets the parent path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> ProjectLocationCatalogBrancheProductImportCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(
self,
name: T,
value: T
) -> ProjectLocationCatalogBrancheProductImportCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

$.xgafv (query-string) - V1 error format.

access_token (query-string) - OAuth access token.

alt (query-string) - Data format for response.

callback (query-string) - JSONP

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

pub fn add_scope<T, S>(
self,
scope: T
) -> ProjectLocationCatalogBrancheProductImportCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.